add field favorite and validateUpdatingFavorite
