+++ routes/api/contacts/validation.js
@@ -1,0 +1,66 @@
+import Joi from "joi";
+import mongoose from "mongoose";
+
+const { Types } = mongoose;
+
+const createSchema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().email().required(),
+  phone: Joi.string().required(),
+  favorite: Joi.boolean().optional(),
+});
+
+const updateSchema = Joi.object({
+  name: Joi.string().optional(),
+  email: Joi.string().email().optional(),
+  phone: Joi.string().optional(),
+  favorite: Joi.boolean().optional(),
+}).or("name", "email", "phone", "age");
+
+const updateFavoriteSchema = Joi.object({
+  favorite: Joi.boolean().required(),
+});
+
+export async function validateId(req, res, next) {
+  if (!Types.ObjectId.isValid(req.params.id)) {
+    return res.status(400).json({ message: "Invalid ObjectId" });
+  }
+  next();
+}
+
+export async function validateCreating(req, res, next) {
+  try {
+    await createSchema.validateAsync(req.body);
+  } catch (err) {
+    return res
+      .status(400)
+      .json({ message: `Field ${err.message.replace(/"/g, "")}` });
+  }
+  next();
+}
+
+export async function validateUpdating(req, res, next) {
+  try {
+    await updateSchema.validateAsync(req.body);
+  } catch (err) {
+    const [{ type }] = err.details;
+    if (type === "object.missing") {
+      return res.status(400).json({ message: "missing fields" });
+    }
+    return res.status(400).json({ message: err.message });
+  }
+  next();
+}
+
+export const validateUpdatingFavorite = async (req, res, next) => {
+  try {
+    await updateFavoriteSchema.validateAsync(req.body);
+  } catch (err) {
+    const [{ type }] = err.details;
+    if (type === "object.missing") {
+      return res.status(400).json({ message: "missing field favorite" });
+    }
+    return res.status(400).json({ message: err.message });
+  }
+  next();
+};
